// src/shared/theme/model/theme-type.ts
export type CustomTheme = {
  colors: {
    point1: string;
    point2: string;
    point3: string;
    textDark: string;
    textWhite: string;
    textMediumDark: string;
    textLightDark: string;
    textExtraLightDark: string;
    textLight: string;
    placeholderColor: string;
    buttonDark: string;
    border: string;

    red: string;
    blue: string;
  };
};